refactor(tests): adopt file names changes
